chore: moved randSec func to config_test.go
